requirements.txt
pandas
pyarrow>=12.0.0
ijson>=3.2.0
boto3>=1.26.0
snowflake-connector-python[pandas]>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=1.4.50,<2.0
Flask-Session==0.4.0
tenacity>=8.2.2

# MySQL connector
pymysql

# Airflow (optional)
apache-airflow==2.7.3
